Take a device type instead of a uuid in initialize()
Currently `initialize()` takes a device uuid, but if we look closer,
there is no need for this. `configure()` requires a uuid since it uses
the unique device identity to generate a `config.json`, but
`initialize()`, in the other hand, just uses the uuid to retrieve it's
device type.

To simplify the interface, `initialize()` now takes a device type slug.